requirements.txt
certifi==2025.8.3
python-dotenv==1.1.1
slack_bolt==1.25.0
slack_sdk==3.36.0
pyobjc-framework-Quartz==10.3.1
